% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.center_predictors}
\alias{.center_predictors}
\title{Internal Helper Function for Mean-Centering}
\usage{
.center_predictors(df, formula, subject_id_col = ".PowR_id")
}
\arguments{
\item{df}{The dataframe to process}

\item{formula}{The lme4 formula to extract predictors}

\item{subject_id_col}{The name of the subject identifier column}
}
\value{
Dataframe with centered predictors
}
\description{
This helper function is responsible for the mean centering of predictors. It detects whether any given predictor
is repeated within clusters (e.g. repeated measurements in participants) and automatically applies within-cluster
mean centering if this is the case. In all other cases, grand mean centering is applied.
}
