% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_data.R
\name{.create_design_matrix}
\alias{.create_design_matrix}
\title{Create the Design Matrix for a Simulation}
\usage{
.create_design_matrix(design, formula = NULL)
}
\arguments{
\item{design}{A \code{PowRPriori_design} object from \code{define_design()}.}

\item{formula}{The lme4-style model formula.}
}
\value{
A tibble (data frame) with predictor variables.
}
\description{
An internal helper function that takes the design specification and an lme4-style formula to generate a
design matrix representing the structure of the experimental design.
}
