% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_data.R
\name{.simulate_outcome}
\alias{.simulate_outcome}
\title{Simulate the Outcome Variable}
\usage{
.simulate_outcome(
  design_df,
  formula,
  fixed_effects,
  sds_random,
  family = "gaussian"
)
}
\arguments{
\item{design_df}{The data frame from \code{.create_design_matrix}.}

\item{formula}{The model formula.}

\item{fixed_effects}{A list of the fixed effects coefficients.}

\item{sds_random}{A list of the random effects' standard deviations and correlations.}

\item{family}{A string indicating the model family.}
}
\value{
The input \code{design_df} with an added column for the outcome variable.
}
\description{
An internal helper function that takes a complete design matrix and simulates
the dependent variable based on the specified fixed and random effects.
}
