% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.PowRPriori}
\alias{summary.PowRPriori}
\title{Summarize a Power Simulation Result}
\usage{
\method{summary}{PowRPriori}(object, ...)
}
\arguments{
\item{object}{An object of class \code{PowRPriori} returned by \code{power_sim()}.}

\item{...}{Additional arguments (not used).}
}
\value{
Prints a formatted summary to the console.
}
\description{
Provides a detailed and context-aware summary of a \code{PowRPriori} object. The output
includes the power table, parameter recovery diagnostics for fixed and random
effects, and (if applicable) calculated Intra-Class Correlations (ICCs). The output is
tailored for different model types (LM, LMM, GLMM).
}
